Last fixes for VOMS Admin 2.5.2
git-svn-id: https://svn.forge.cnaf.infn.it/svn/voms-admin/voms-admin/branches/licensing_work@219 65f5e8a5-84a7-4ea7-892f-e18654e157d3

+++ resources/scripts/voms.py
@@ -134,7 +134,9 @@ def backup_dir_contents(dir):
     
     ## Remove backup filez
     for f in backup_filez:
-        os.remove(f)
+        ## Don't remove backup directories potentially created by the user
+        if not os.path.isdir(f):
+            os.remove(f)
     
     filez = glob.glob(os.path.join(dir,"*"))
     backup_date = time.strftime("%d-%m-%Y_%H-%M-%S",time.gmtime())
